Refactoring. Done a nested class Node in MyLinkedList
