+++ AShchepets/src/main/java/week3/linkedList/MyLinkedList.java
@@ -241,4 +241,46 @@ public class MyLinkedList implements MyList {
         }
         return false;
     }
+
+    // ********* NESTED CLASS NODE ***********
+    public static class Node {
+
+        private Object data;
+        private Node next;
+        private Node prev;
+
+        public Node(Object data) {
+            this.data = data;
+        }
+
+        public Node(Object data, Node next, Node prev) {
+            this.data = data;
+            this.next = next;
+            this.prev = prev;
+        }
+
+        public Object getData() {
+            return data;
+        }
+
+        public void setData(Object data) {
+            this.data = data;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+    }
 }
